feat(deploy_kube_feature): deploy to platform
