+++ packages/provider/src/hooks/use-context.ts
@@ -3,6 +3,6 @@ import { useContext } from 'react';
 import { CoreComponentsContext } from '..';
 
 /**
- * общий стор
+ * Общее хранилеще настроек для всех вложенных компонентов
  */
 export const useCoreComponentsContext = () => useContext(CoreComponentsContext);
